Minor: do not hardcode DefaultQueryTransformer sequence.
(cherry picked from commit 447c9c27aacbc1524a70da195096e8954c03accf)
